requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
black==23.12.1
Brotli==1.1.0
click==8.1.7
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
isort==5.13.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
yarl==1.9.4
